Split local server route modules
